separate generic data (JSON) module + service socket communication will established only, if the module is loaded
