+++ web-app/src/app/app.reducers.ts
@@ -2,7 +2,6 @@ import { ActionReducerMap } from '@ngrx/store';
 
 import * as fromCore from './core/store/core.reducers';
 import * as fromDataTypes from './core/datatypes/store/data-types.reducers';
-import * as fromGenericData from './eep/generic-data/store/generic-data.reducers';
 import * as fromSignal from './eep/signals/store/signal.reducers';
 import * as fromTrain from './eep/trains/store/train.reducer';
 import * as fromStatistics from './eep/statistics/store/statistics.reducer';
@@ -10,7 +9,6 @@ import * as fromStatistics from './eep/statistics/store/statistics.reducer';
 export interface State {
   core: fromCore.State;
   dataTypes: fromDataTypes.State;
-  genericData: fromGenericData.State;
   signal: fromSignal.State;
   train: fromTrain.State;
   statistics: fromStatistics.State;
@@ -19,7 +17,6 @@ export interface State {
 export const reducers: ActionReducerMap<State> = {
   core: fromCore.reducer,
   dataTypes: fromDataTypes.reducer,
-  genericData: fromGenericData.reducer,
   signal: fromSignal.reducer,
   train: fromTrain.reducer,
   statistics: fromStatistics.reducer,
