+++ web-app/src/app/client/modules/json-data/generic-data-overview/generic-data-overview.component.html
