+++ web-app/src/app/client/modules/json-data/generic-data/generic-data.component.html
