+++ web-app/src/app/client/modules/json-data/store/generic-data.actions.ts
