--- web-app/src/app/eep/generic-data/model/generic-data.ts
@@ -1,12 +1,0 @@
-// export class GenericData {
-//   public constructor(public val1: any,
-//                      public val2: any,
-//                      public val3: any,
-//                      public val4: any,
-//                      public val5: any,
-//                      public val6: any,
-//                      public val7: any,
-//                      public val8: any,
-//                      public val9: any,
-//                      public val10: any) {}
-// }
